Sections: fixed hardcoded key transfer with actualy key
There is still a hardcoded fingerprint of the key we want it signed, this is because we haven't yet implemented the scanner. For now to test the key transfer you should set SCAN_FINGERPRINT to the fingerprint of the key you want it scanned (the one you select it).
Open two Geysign instances, in one select a key and press 'Next' until you've reached QR code page, on the other one go on 'Get Key' Tab and click 'Download Key'. There is still a fix to be made because for now the key will only be transferred when you stop the keyserver (in the instance of Geysign where you selected the key, click on 'Back' button).