Keyserver: Made it bind to IPv4 *and* IPv6 socket
The ThreadedKeyserver inherits from BaseHTTPServer which itself is a
TCPServer.  It uses a bind_server method to allocate the socket.
Apparently there is a IPV6_V6ONLY option which determines whether
a IPv4 socket should also be bound.

According to very simple testing, it does indeed bind to both,
IPv6 and IPv4 addresses.