SignPages: Not showing the actual signatures by default
The code to show the signatures has not yet been deleted.
You can set SHOW_SIGNATURES to True and it will print the last 10
signatures.  However, I don't like the current design.  It might be
usefuly to have and in fact our "content" is signature, so we should
somehow present those.  But how many and which should be present?  And
what do we offer to do with those?  The user cannot really "remove" the
signatures or edit them in any way, anyway.  So we just do not show the
signatures here, now.