Added a Logo
This is a bit preliminary though and was done very quickly.
It uses elements from Seahorse GPLv2+ licensed.
It also uses element from the GnuPG logo, downloaded from
http://upload.wikimedia.org/wikipedia/commons/a/a3/GnuPG-Logo.svg
and, according to the corresponding page:
http://commons.wikimedia.org/wiki/File:GnuPG-Logo.svg
is in the public domain.
It also uses the GNOME logo obtained from
http://commons.wikimedia.org/wiki/File:GnuPG-Logo.svg
licensed under the LGPLv2+.