Added a .desktop file
One problem I encountered was that gnome-shell (rightfully) wants to
be able to exec() the program in the Exec line.
That, however, requires either the binary to be in the path or
an absolute path to the binary.
The former is hard to achieve as ~/.local/bin/ does not seem to be
in the PATH by default.
The latter is also a bit hard as I don't know how to make setuptools
rewrite the Exec line with the prefix it installed to.
Now, we simply call python with the module's name.
This is a bit of a hack but should work when the user installs the
application via pip, i.e. pip install --user. As a bonus, this
should work equally well for system wide installations.
Installing to virtualenvs is not supported in the sense that the .desktop
file won't show up in GNOME Shell automatically. One would have to
convince GNOME Shell first to look into the new directory for the .desktop
file and then convince python to find the module.  But I consider that
to be an esoteric case for now.