Sections: made verify_fingerprint return no whitespace
It was cumbersome to copy and paste the fingerprint from one window to
another, because it contains a linebreak (\n).  This linebreak was used
for verification of the received fingerprint.  That, however, does not
contain any linebreaks...  This commit changes the behaviour such that
the fingerprint entered will be cleaned from whitespaces.