MainWindow: Added documentation for the setup_server function
Actually, this is a bit of a weird design.  setup_server is being called
from Sections.py which got passed a reference to the "application".  I
think we can do better by throwing signals when the user selected a key.
Then we can catch that signal and start the server accordingly.