TmpKeyringCopy: Copying all public keys to the tmp keyring
Instead of only those which are usable.
The rationale is that we need to responding public keys to decide
whether a secret key is usable.

We could, alternatively, somehow remove the unusable secret keys, but we
didn't want to copy the secret keys around in first place. And I don't
think it's possible to make certain keys not appear.