scan_barcode:  Enabled image for barcode on legacy systems
In fact, the patched GStreamer zbar plugin hasn't even hit the Git
repository yet.  So it'll take a while before users will have it.  In
order to have an "interim" solution, we use the workaround of listening
to every frame ever posted on the bus.  This is expensive.  There may be
other, smarter, ways of achieving the desired result.