scan_barcode: Preparing for new GStreamer
I'm trying to build something which can be used by both, the newest
GStreamer with the zbar element having the attach-frame property and
older, actually deployed GStreamer versions.  This is an approach which
seems to work, but it quite slow.  I probably needs to be replaced by
something more sane.