SingKeyPage: Display the scanned image
When the user scanned a QR code, we display it in the "review" tab.
This is to prevent an attack which might be considered exotic: An
attacker injecting a malicious image for a frame or two, just enough for
a barcode to be decoded.  The user shall be enable to review what has
been scanned to then be able to cancel the operation.  Another extension
to this could be to highlight which part of the image contains the
detected and decoded barcode.