scan_barcode: Using a simpler pipeline for obtaining the barcode image
This new approach is hopefully a little bit simpler and easier to
understand, although it is multistaged.  That is, we require the barcode
element to post its message on the bus before the pixbuf element does.
Also, the pixbuf element should post with the very same timestamp as the
zbar element.
This might very well be a bit more expensive because the pixbufs are
created all the time.