GetKey: When receiving the key, import into a new keyring
rather than re-using the existing one.
This is not optimal in the sense that we might end up importing twice:
once for the verification and the second time when receiving.
But re-using might lead to subtle bugs such as when the client tries to
download a key and fails to verify a key. Currently, verification
imports the key but does not delete them. And it also uses a construct
like tmpkeyring.get_keys().keys()[0] which looks like it could be broken
easily by adding a second key.