GetKey: avoid lists when interfacing with gpg
Currently, monkeysign's gpg module returns dicts when it's being asked
for keys.  The dict, in python2, likes to return lists. But in python3
this won't be the case. Let's try to be a little clever about it by
using (arguably ugly)   next(iter(dict.items()))  instead of
dict.items()[0].  Unless we really expect only one key.  So we're
actually not using that mentioned construction at all...