KeyPresent: Relatively importing packages
This includes a crude hack for running the file directly,
i.e. python keysign/KeyPresent.py. That is, if I read PEP 338 correctly,
discouraged. However, I'm an old fart and it seems as if the proper way
requires me to install an executable via distutils or call with
python -m keysign.KeyPresent.  The former is a non-starter for my
development routine. I tend to run the scripts often while editing. I
also dislike the idea of installing a new executable onto the users'
machines.  I dislike the latter mainly because my shell doesn't auto
complete as nicely.  Both are probably not very good reasons, but I'm
happy to carry these 10 lines of code around to make my life easier.