GetKeySection: Parse fragments out of the barcode URL
So far, we've encoded the MAC as query part in the URL. This can be done
a little bit better by using the fragment part which does not get sent
by the HTTP client.  We prepare for this upcoming change by parsing the
fragment part in addition to the query.

There is a potential problem by mangling the query and the fragment: We
could overwrite existing values.  As we're only making use of one value
right now (MAC), we should be fine though.