KeysPage: Removed displaying keyid/fingerprint
Actually we don't need it, do we?  Currently, the (synthesized) keyid is
displayed in the right pane and the fingerprint on the next page. If we
assume the user to have different keys (i.e. fingerprints) with
different UIDs, then the keyid as distinguisher does not make much
sense. Of course, users might indeed have multiple keys for the same
UIDs. Then, we hope the information displayed when selecting a key is
enough.