gpgmh: Added Key and UID classes
The main reason is to be able to replace the gpg providing library more
easliy as we intend to switch to something gpgme based.

Subclassing namedtuple gives us a nice repr and a little bit of sanity
check when creating new objects.

We're requiring the expiry for now as we're displaying the expiry of a
"key" in one GUI element. We could have required the Key to only carry
an expiry flag, but to make the interface more similar, we also require
an expiry when creating a UID.  This may very well change.

The classes may also get additional (synthetic) properties such as
"expired". But we'll add them once we need them. Again, our goal is not
the build a comprehensive and complete representation of a key but only
of the elements that we need.