KeySignSection: Remove switch_to_key_present_page function
There is only one place calling that function. In that scope, we have
the keydata which we need for the MAC. By removing the function, we
eliminate the need to pass that data with the function call. And the
removed function does three things, only which we can include in the
caller function.