KeySignSection: Calling construct_key_present_page with qrcodedata
Now we know that we produce the MAC for the very same data that we pass
to the Keyserver, because it is being set up just a few lines above.
This is better than acquiring the keydata later, because we are not
racy. Also, the code should be more comprehendable if operation with the
keydata happen in one place rather than many.