setup.py: added keysign to package_dir and added appdata.xml
So far, the appdata.xml file had not been installed.
Adding keysign to package_dir shouldn't be required for now, but if we
add .ui files later, we might very well require it to be set.  Also, we
have the 'keysign' package in that directory, so let's mention it
already.