send.ui: Added an InfoBar
For showing when the user does not have any keys.

This patch looks large. Partly because of a lot of whitespace changes
(use -w to ignore those), but also because there are other random
changes in the UI file which accumulated over time :-/  I should have
made more commits on the way.