send: stop avahi offer first
This UI gets stuck for a few milliseconds when switching back. It seems
that shutting down the HTTPd is the problem.  I tried many things.
Giving control back to the mainloop with an experimental generator based
wrapper, spawning Threads, deferring for a few milliseconds. It's all
not very nice.  The real fix has yet to be found.