Keyserver: Defined the poll_interval to be 0.15 instead of the default 0.5
The HTTP Server is a BaseServer which does has following in its
serve_forever event loop:

            while not self.__shutdown_request:
                # XXX: Consider using another file descriptor or
                # connecting to the socket to wake this up instead of
                # polling. Polling reduces our responsiveness to a
                # shutdown request and wastes cpu at all other times.
                r, w, e = _eintr_retry(select.select, [self], [], [],
                                       poll_interval)

This causes the UI to stall for max 0.5 seconds.  This induces a noticable delay.
I hope that 0.15 does not waste overly much CPU.  Quickly glancing at "top"
indicates that it doesn't leave any visible traces...

Another option might be to connect to that socket (from another thread)
just after having requested the shutdown. But this doesn't work so well
for me. I guess due to Python's GIL.

Some day, we might want to use a GLib-based server.
I'm not aware of any existing implementation.