SignPages: Made it accept a key and an image in the constructor
We want to have immutable widgets as much as possible, because I believe
it makes it easier to work with the widgets. I.e. you only need to know
what to pass them during creation instead of learning the functions they
provide.  This easy modification makes it require a key to be given.