GetKeySection:  Use the new PreSignWidget conditionally
When setting the KEYSIGN_NEWUI environment variable, the new widget will
be shown.

We're doing it conditionally, because we don't want to expose the new UI
just yet.  I'm missing some more polish.  I intend to enable users to
play around with the new UI and provide feedback.

Unfortunately, this removes the option to display an error :-/