gpgmh: Take homedir not keyring in get_usable_keys and get_usable_secret_keys
We don't expect the user of the library to know about a keyring, because
currently, it's a Monkeysign thing and we intend to be able to use other
libraries, too.

This also includes a refactor of a get_usable_keys_from_keyring
function.  It deals with a Monkeysign Keyring, gracefully handles no
results and filters for usable keys.