scan_barcode: Allow widget to be run with GTK CSD
It seems there is a bug with GTK and GStreamer:
https://bugzilla.gnome.org/show_bug.cgi?id=721148#c6

When using client side decorations, it breaks gstreamer displaying
stuff. This is unnoticed as long as no CSDs are used. But when wanting
to use a GtkHeaderBar, CSDs are used and thus the widget breaks.

To force CSD, you can run this program with GTK_CSD=1, i.e.
env GTK_CSD=1 python keysign/scan_barcode.py