added an example receive application
The counterpart to the send application which receives and signs a key.

It's more of a proof of concept rather than an app that should be used.
Ideally, we get to the point where we can have the two separate apps and
somehow simply merge them to a bigger app having both the send and
receive part.