Added a combined "app" application
This should become the future application.

It's a bit grown and needs to be weeded out. But it works reasonably
well for now.  I wish it was more modular in the sense that the relevant
bits and pieces of the send and receive parts were more separate and not
copy and pasted from the send and receive application.