avahidiscovery: Added a MAC checking class
One of our features is to download an authentic copy of the other
party's key.  We use a MAC to assert that both sides have the same key.
Currently, the MAC is transmitted securely via the barcode and only via
the barcode.  We allow, however, manual input of the fingerprint. That
does not carry MAC information.  So we try to opportunisitically match
the MAC.

This will bite us one day, so we better get rid of the manual
fingerprint method soon.