receive.ui: Keyconfirm widget: Default the confirm button to have focus
Weirdly enough, the glade file says that the labels are not able to get
focus, but Builder creates them as being able to have focus.  This is a
bit annoying because when the label has focus (as first widget...) then
it paints the background as if everything is selected which looks just
weird.