send.ui: Reparented the keylist content into a separate box
to have the infobar below the stack itself.  The main motiviation is to
apply some margin to the actual keylist. With the old widget hierarchie
the box containing the infobar would have needed some margin.  I think
the infobar looks better when it's across the whole widget, i.e. without
margin.