receive: Adapt to the new stack with the scanner widget child in a box
We shuffled the scanner around a bit. Now we need to adapt every
consumer of the scanenr :-/

When assembling the UI we have a ScannerWidget created. That basically
returns newly created widgets. We need to add them to our window,
somehow.  The point of injection of the new scanner is determined from
the "scanner_widget" in the UI file. Because we have a new scanner, we
can (and need to) remove it from the Builder and then add the new
scanner.

It's still a bit fragile.