gpgmh: Use the TempSigningKeyring with a base_keyring
Admitted, the API is sufficiently weird.  But we're getting rid of it,
anyway.

The reason for this change is to allow a test to specify a homedir, i.e.
to not leak anything about the regular user's keyring.
Before, the user's keys were visible, because the TempSigningKeyring
copies the secret keys from a freh Keyring() is no base_keyring is
provided. That fresh Keyring uses the user's regular keys.  We provide
the homedir, because the TempSigningKeyring calls Keyring's __init__
which sets the keyrings' homedir accordingly.