gpgmh: sign_keydata: raise if no error_cb was given
We might miss very subtle bugs if we ignore by default.  We actually had
a problem with the homedir not being set properly for a
TempSiningKeyring which was much eaier to debug with the function
raising by default.  I'm also questioning the concept of an error_cb
rather than catching exceptions.  The callback has the advantage of not
interrupting the loop if that is desired.  I don't think you can
continue the loop when you throw.