scan_barcode: Removed BarcodeReader object
And moved everything to BarcodeReaderGTK.

This allows us to port to the GStreamer's GTK sink more easily, because
we do not have to care about the non-GTK version of the scanner.

The diff looks bigger than it really is...

The prepare-window-handle message was handled in the BarcodeReaderGTK
only and has been integrated into the BarcodeReader's on_message
function.  The run method has been moved to the GTK class.