BarcodeReader: Moved from xvimagesink to gtksink
With the gtksink, we don't need to handle prepare-window-handle any
more.  So we can get rid of the XWindowID alltogether.

The BarcodeReaderGTK also becomes a Box now instead of a DrawingArea,
because the gtksink provides us with a proper GtkWidget which we add to
a container of our choice (which is a Box for now).