BarcodeReaderGTK: Switch from v4l2src to autovideosrc
I think this semantically more correct, because we don't insist on v4l2
being present (think non-Linux).  Also, I was told that it should pick
up Pinos, the (new) Webcam-multiplexer which may be what will be used in
Flatpak-land.

As a bonus, we add a line to activate loading a video from disk, if you
have recorded one.