scan_barcode: Trying to fix up the ScalingImage requesting its size
For an unknown reason, the gtksink acquires space relatively
aggressively and doesn't allow for the ScalingImage to be rendered.
Unfortunately, I know very little about Gtk and how to deal with a
DrawingArea. For now, I hope for a widget from Gtk directly that can
show me an image as big as possible.