receive: named box20 "scanner" to switch back to it easier
We had the problem of not being able to reliably switch back to the
scanner page.  It works well if we know the container of the scanner but
we don't necessarily do anymore ever since we do not add the scanner
directly to the stack but rather a container which also holds an
InfoBar.  Now we gave the stack's immediate child a name: "scanner"
which we can use to switch back to.