README: Re-introduce python-gi-cairo as dependency
We need the (very light) GI bindings for drawing the barcode.
I think we need the gi-cairo because the on_draw method exposes some
cairo data structures. Because on_draw is within Gtk, which itself is
GI, we need the cairo-gi bindings. The python-cairo is, in fact,
a dependency of python-gi-cairo (at least on Ubuntu).

Reverts: 214a8f9fbbccc519eda6962d7e59df62e12f2680