scan_barcode: Use pack_start instead of add
We used the generic add() to attach the Gst widget to ourselves. We
think we are a Box.  A Box has, in addition to add(), a more precise
pack_start() function.  Funnily enough, we crash here in certain
circumstances.  I haven't understood the full details, but it works when
executing keysign.app, i.e. python -m keysign.app, but not when
executing the keysign module directly, i.e. python -m keysign.
Even more funny:  It doesn't crash when launching from the repository,
but not when launching from outside the repository, i.e. home directory.