gpgmh: Use a local logger in UIDExport
We get a local logger instead of the module level one to allow
configuring of the loggers.  The main application may very well not have
configured the logging system before this library is imported.  This
should be the standard case, actually.  In that case an unconfigured
module level logger exists.