gpgmh: Remove the export-minimal option in sign_keydata
This solves a long standing annoying bug.  The keys sent would not carry
new signatures.  The reason seems to be that we're exporting minimally.
That, by definition, strips everything but self-signatures.

We don't need to export minimally, because a few lines further down we
explicitely request a minimal export of the key.

I don't know why this option was introduced in first place.  But from
briefly testing it seems to work way better.