receive.ui: Use a Box instead of a Frame for the scanner
To fix a very annoying crash.

I don't fully understand why it crashes with a frame, but it seems to
lose the object. It crashes with a segfault. Probably an objects being
free()ed but re-used.  Maybe the GtkFrame is a bit special regarding
handling children.

A GtkFrame is the wrong widget, anyway.  We probably thought that it is
a relatively simple and lightweight container which provides space for
exactly one child.  But that is not the case.  We are using a Box now
although all we need is a Bin.

With this change, the crash is gone and the application runs happily.