gpgmks: Refactored gpgmh to a gpgmks module
It should be a little more clear now. We have gpgmh which loads either
gpgmeh or gpgmks. It loads the gpgme-based library only if the use
requested it by setting the KEYSIGN_GPGME environment variable.  We
will, in the near future, load the gpgme-based wrapper though and only
fall back if that's not available.