setup: Build and ship translations
It seems unneccessarily hard to do i18n.  I think we want to have the
translations compiled when compiling the Python stuff rather than just
before we're installing stuff in some prefix.  Hence we're adding the
compilation of the PO files to the build step rather than the
installation step.

The decision to put the MOs as package_data rather than data_files has
been made because it seems to be hard to do.  data_files does not seem
to take wildcards. How would we dynamically determine the list of mo
files to be installed? And how would each mo file be installed in their
respective directory? And how can we make it work for pip --user
installs off a git URL?  With the package_data it seems to work well
enough for now.