Made all Python files executable
It seems to upset rpmlint to have files with a shebang.  The shebang
doesn't make much sense if the file is not marked as executable. Now you
could also remove the shebang, but then some editors (most importantly:
mine) uses the shebang to detect what file type is being edited. To make
both me and rpmlint happy, we now mark the files as executable.