keylistwidget: Default to empty string for UID components
We cannot use the "bytes" constructor for extracting the values there,
because in Python 3 it needs an encoding when given a string. In Python
2, however, it does not take an encoding argument.
All it wants to achieve, anyway, is to limit the number of items that
are treated with the re-encoding procedure later (d = ...).  So we
really only want to cut out a piece of the UID instead of touching all
values of the UID.  We also want all components to be coerced into a
string.  We expect the name and email address to be a string.  The
expiry is an int or a datetime.  Hence we try to concatenate with a
string in order to coerce the value.  We default to the empty string in
order to treat None.