send tab: refactor widget signals handling
Instead of creating a KeyPresentWidget with no informations when a key
has been selected, I delayed it directly when we have the informations
about the code to display in the KeyPresentWidget.
Because of this change we cannot connect to kpw signals in on_activate
of KeysignApp, because when on_activate is executed we have not yet
initalized kpw.
So instead now we react to the visible-child signal.
Also with this approach a bug about "map" not being called when we
reached the result page and pressed the back button, that appeared with
the commit 51d6582, has been fixed.