gpgkey: Make the UID "safe" for direct consumption
If we have a key with a non-UTF-8 UID gpgme will return a proper string
with surrogates.  These are the actual undecodable bytes encoded in a
way that allows to get hold of the actual undecodable bytes later when
we need them.  We don't though, because we don't have to call gpg with
the exact byte sequence again.  Hence we can make the UID safe for
display centrally, i.e. in the gpgkey.UID class.