gpgmks: sign_keydata_and_encrypt: Use the fingerprint as recipient
Rather than the UID.
We've modified the UID to be a consumable string.  As such, we will not
be able to use it as an identifier for gpg to identify a key. We could
somehow pass the raw UID bytes around, but using the fingerprint seems
like much less effort with very similar results.