setup.cfg: Remove the where clause
I haven't understood yet how to make use of that properly.  It seems
that python setup.py test works regardless. Probably not as nicely,
because it'll scan all directories for tests, but it confuses nose if we
leave that entry in, e.g. running specific tests seems to be much
harder.