tests: gpgmeh: Make it compatible with gpgme 1.8 and gpgme 1.9
It seems that newer versions of the gpgme python bindings set the mode
when calling keylist().  This is nice, but is not possible to do in 1.8.
1.9 also overwrites whatever you set before with set_keylist_mode.
We now try to be compatible with both.
Eventually, once we can safely ignore gpgme 1.8, we can get rid of the
compatibility.
Currently, Fedora ships gpgme 1.8.