setup.py: Generate translated desktop file
This seems to work reasonably well for having a translated desktop file
generated.  I don't like so much having all of this in setup.py so maybe
someone gets around to extract it into a babel-desktop package or so.
But it's the first step towards having a useful babel setup for desktop
apps.