gpgmeh: Let gpgconf determin the homedir
gpgconf does the right thing, i.e. respecting the GNUPGHOME environment
variable, so we can as well let it do its job.

I got into trouble when using a second instance with such an environment
variable. The logic would revert my efforts by manually overwriting the
homedir with my actual homedir. That is not desireable, of course.
This change makes it work as expected.