gpgmeh: Filter offline keys
Debian recommends placing the master key offline. That requires the
master key to be present for signing other keys.
This change filters those keys whose master key is not present.
In the future we could make our UI a bit nicer by showing the key
somehow in a disabled fashion. We could offer a hint that the user needs
to provide their master key.  But we cannot do it just right now,
because we need to drop monkeysign support first.