desktop file: run with python3
This is a cheap way to get the desktop file working in Flatpak again.
We used to run Python 2 in the sandbox. Now we have bumped that to
Python 3 even though have don't really have the need. The app is still
bilingual.
In flatpak, however, all dependencies are installed for python3 and
Flatpak's python is python2. So the app doesn't start.

Now we could make the desktop file execute "gnome-keysign", but that
requires that name to be in the PATH. That is a problem for casual
installations via pip, because ~/.local/bin/ is not in the PATH by
default on many distributions.

The other option is to use a dedicated .desktop file for the Flatpak
build.

This change is the easiest to make, albeit being not most correct. That
is, the app does not require python3. But most people should be running
python3, anyway...