Replace python-avahi with a vendorized version
python-avahi is a package that just provides some constants. The problem
we were facing is that Debian is currently providing only python2-avahi
and not pytho3-avahi
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=853239

With this commit we simply ship ourselves the Avahi constants we need,
like Gajim already did
https://dev.gajim.org/gajim/gajim/merge_requests/255

In AvahiBrowser there were also two unused imports (DBusException and
Gio), so I used this opportunity to remove them.