Re add the "txt_array_to_dict" check
This function is not yet available with Avahi. Even if with our
vendorized Avahi now we were sure that it was always there, keeping this
check in place will ease our eventual transition again to python-avahi.

The special check regarding the avahi versions below 0.6.14 has been
removed because that version is now 12 years old (2006). So we should be
safe enough even without it.