discover: fix error if discover was stopped
If a user typed really fast in the receive tab, self.stopped was set at
True and the return of the "start()" function contained an empty
message.
Then in receive.py with "_receive()" the message was propagated reaching
the user interface as if an error occurred.

With this fix we return the LonelyError that is already gracefully
handled in "_receive()".
LonelyError fits this scenario because it indicates a connection stopped
before a transfer.