delete gpgmh
This module is not necessary any more, because we use gpgme exclusively.
  We used to have this abstraction, because we were using monkeysign and
unify gpg access through this module.  But now we have gotten rid of
monkeysign and there is no point in having a wrapper for one underlying
library.  So we use gpgmeh directly.  This is a fairly mechanical
change, i.e. sed -i s,gpgmh,gpgmeh,g.  It seems to work well, though.