wormholeoffer: prevent unhandled LonelyError
With the recent wormhole versions (probably from the 0.11.0) the close()
behaviour slightly changed.
In keysign output we got this message when we cancelled a pending
sending operation:
```
Unhandled error in Deferred:

Traceback (most recent call last):
Failure: wormhole.errors.LonelyError:
```
To solve this I added an error callback to the close function. Because
we still catch every possible errors in the start() function, here I
just redirect the error output to `log.debug`.