tests: Fix key import error
Apparently we had another problem with these tests.
Calling get_public_key_data with the required temporary homedir worked
as expected.
The problem was that in wormholeoffer, bluetoothoffer and avahioffer we
also call get_public_key_data, but without the temporary homedir that
has the private key we were trying to use.
With a "GNUPGHOME" environment variable we solve this problem.