sign-key: Actually print to stdout that we are waiting for the user
It seems silly to have logged that statement. If only because the log
level can cause the message to not be printed at all. And that is what
happened.  Now we use print() which should certainly make the message
land on the user's screen.  We could think about using stderr, but let's
not overengineer this for now.