gpgmeh: remove detection for crashy gpgme
We used to work around a bug on 32bit platforms which crashed the
process. It's been a while and we may as well assume that the issue has
been fixed for good.

This neatly solves a problem of the gpgme version returning not only
ints, but also strings. We don't handle that case very well.