Added split UIDs programme
You may want to split a key up into various files each containing a UID
of yours. This small tool helps you to achieve that. It's quite rough,
e.g. it does not yet have a nice --help output and the minimise option
cannot be given during invocation. That might change in the future,
though.