gpgmeh: raise error if we import a new certificate rather than new sigs
The attack is subtle and maybe not very relevant:
Two parties have exchanged their certificates and now one side is
waiting for the email with the newly produced certifications (aka
"signatures"). The attacker sends a prepared email with a new
certificate that it wants the victim to import. Maybe to poison the
keyring or to make the victim look bad by placing some phishy
certificates.

This change attempts to detect that and raises an error if it thinks it
is being tricked.