send: split the import infobar up in to two infobars
Now we have an "okay" and an "error" case. The reason is to manage less
state with a future "return signature" button. Such a button would need
to disappear in the error case (and re-appear in the happy case).
Instead of juggling with the visibility state, we make two infobars that
can be made visible.