tests: move to pytest
The timeout for tests involving Twisted seems to have disappeared. Let's
hope it doesn't cause trouble when the tests will, in the future, fail.

The nose style assertion functions are obsolete with pytest. It
manipulates the bytecode to inspect the arguments of the assertions, so
that one gets a nice enough output when things fail.