setup: replace nose with pytest
It seems to be discouraged to specify a test runner in setup.py. I
haven't fully understood why, but it seems to be related to the
dependencies being pulled. Or so.

Since I don't know what the Python equivalent of "make test" is, I keep
providing a test command. You are still free to not use that.

The tests run with a simple invocation of "pytest". If you want to see
the logging, "pytest --log-cli-level=DEBUG" will do. To test a specific
test, run "pytest tests/test_bluetooth.py::test_bt_wrong_hmac"