tox: define pytest environments
I couldn't get it to run. It seems to require the app to be built as
wheel which fails due to the appdata file not being present.
I guess it's a bug somewhere, but so far, nobody complained.
I also don't know whether it is a bug in the first place nor how to fix
it.