tests: move get_signatures_for_uids_on_key to gpgmeh
We need this function for finding out who sent us a certification.
Because we want to know the lits of signatures before importing the key
and compare it to after the import. Well. Ideally, we could analyse the
certification structure and determine the producer of the signature. But
gpgme does not allow for that. We could probably depend on a parser,
maybe Sequoia, but I don't know how well it would work.