gpgmeh: make get_signatures_for_uids_on_key return attestors
So we're hiding gpgme structures from the caller. Currently, this is all
we need, so we might get away with it. I have a mild concern for the
type of signatures, i.e. currently we are including each and every type
and not only certifications. I hope it won't cause a problem down the
line.