send: detect who produced the certification when importing
The goal is to be able to send the received certification back to where
it came from. Because.. well. the certification is not worth so much in
our hands and has much more value to the one who produced it. But we're
not having a channel back to the sender. So we're trying to find ways to
identify who produced the certification by looking into the keyid field
of the newly imported certification. It's not very precise as it
identifies keys, not email addresses.