send: actually return the certifications to the sender if the key is known
We know which (short) key id has produced the certification.
We do not know the sending address, though.
But if the attestee has the key of the attestor, we can infer which
email addresses we can send the certifications back to.

The actual button in the info bar still shows "undo" which is bad, of
course.
And I am not happy with the instruction we give to import the
certifications, but we can take it from here and improve the text
gradually.