receive: allow for dragging a pubkey file or contents for certification
If we ever happen to have received or downloaded a file with a public
key, we want to be able to create a certification much like with the
signtool.

We may run into a problem, though: We also allow for dragging the
received certifications into the app. The use case is slightly different
and we may have trouble explaining to users that dragging the file in
one area works but does not in another.