Avoid SyntaxWarning with Python 3.12
Python 3.12 warns about invalid escape sequences.

'[\s]+' contains an invalid escape sequence.  Python just treats it as
a literal backslash followed by an s, but (for example) '[\n]+' would
behave differently.

We fix this by telling python to treat the string as "raw" -- that is,
not to escape the sequence at all.  This lets the regular expressions
module see the backslash directly, as intended.